+++ mir_commander/core/logging.py
@@ -36,6 +36,8 @@ def setup():
     )
     logging.getLogger("cclib").setLevel(logging.CRITICAL)
     logging.getLogger("OpenGL").setLevel(logging.WARNING)
+    if FROZEN:
+        logging.getLogger("httpx").setLevel(logging.WARNING)
 
 
 __all__ = ["setup", "getLogger"]