+++ mir_commander/ui/settings/settings_dialog.py
@@ -38,7 +38,7 @@ class SettingsDialog(QDialog):
         app_config: AppConfig,
         project_config: ProjectConfig,
     ):
-        super().__init__(parent)
+        super().__init__(parent, modal=True)
 
         self.app_apply_callbacks = app_apply_callbacks
         self.mw_apply_callbacks = mw_apply_callbacks
@@ -106,8 +106,8 @@ class SettingsDialog(QDialog):
         """Generation of particular pages (as tab widgets) with controls for settings."""
 
         page_items: list[dict[str, Any]] = [
-            {"title": self.tr("Project"), "tabs": [(Project, "")]},
             {"title": self.tr("General"), "tabs": [(General, "")]},
+            {"title": self.tr("Project"), "tabs": [(Project, "")]},
             {"title": self.tr("Plugins"), "tabs": [(Plugins, "")]},
         ]