+++ resources/i18n/ru.ts
@@ -4,54 +4,122 @@
 <context>
     <name>About</name>
     <message>
-        <location filename="../../mir_commander/ui/about.py" line="41"/>
+        <location filename="../../mir_commander/ui/about.py" line="43"/>
         <source>Version {}</source>
         <translation>Версия {}</translation>
     </message>
     <message>
-        <location filename="../../mir_commander/ui/about.py" line="48"/>
+        <location filename="../../mir_commander/ui/about.py" line="50"/>
         <source>About</source>
         <translation>О программе</translation>
     </message>
     <message>
-        <location filename="../../mir_commander/ui/about.py" line="49"/>
+        <location filename="../../mir_commander/ui/about.py" line="51"/>
         <source>Authors</source>
         <translation>Авторы</translation>
     </message>
     <message>
-        <location filename="../../mir_commander/ui/about.py" line="50"/>
+        <location filename="../../mir_commander/ui/about.py" line="52"/>
         <source>License</source>
         <translation>Лицензия</translation>
     </message>
     <message>
-        <location filename="../../mir_commander/ui/about.py" line="59"/>
+        <location filename="../../mir_commander/ui/about.py" line="61"/>
         <source>A modern, powerful graphical user interface for molecular structure modeling.</source>
         <translation>Современный, мощный графический интерфейс для моделирования молекулярных структур.</translation>
     </message>
     <message>
-        <location filename="../../mir_commander/ui/about.py" line="66"/>
+        <location filename="../../mir_commander/ui/about.py" line="68"/>
         <source>Home Page:</source>
         <translation>Домашняя страница:</translation>
     </message>
     <message>
-        <location filename="../../mir_commander/ui/about.py" line="68"/>
-        <source>Telegram News:</source>
-        <translation>Новости в telegram:</translation>
-    </message>
-    <message>
-        <location filename="../../mir_commander/ui/about.py" line="80"/>
+        <location filename="../../mir_commander/ui/about.py" line="82"/>
         <source>Yury V. Vishnevskiy</source>
         <translation>Юрий Вишневский</translation>
     </message>
     <message>
-        <location filename="../../mir_commander/ui/about.py" line="82"/>
+        <location filename="../../mir_commander/ui/about.py" line="84"/>
         <source>Valery V. Vishnevskiy</source>
         <translation>Валерий Вишневский</translation>
     </message>
     <message>
-        <location filename="../../mir_commander/ui/about.py" line="28"/>
+        <location filename="../../mir_commander/ui/about.py" line="30"/>
         <source>About Mir Commander</source>
         <translation>О Mir Commander</translation>
+    </message>
+</context>
+<context>
+    <name>ApplicationUpdateDialog</name>
+    <message>
+        <location filename="../../mir_commander/ui/updates.py" line="87"/>
+        <source>Ok</source>
+        <translation>Ок</translation>
+    </message>
+    <message>
+        <location filename="../../mir_commander/ui/updates.py" line="89"/>
+        <source>Cancel</source>
+        <translation>Отмена</translation>
+    </message>
+    <message>
+        <location filename="../../mir_commander/ui/updates.py" line="91"/>
+        <source>Skip this version</source>
+        <translation>Пропустить эту версию</translation>
+    </message>
+    <message>
+        <location filename="../../mir_commander/ui/updates.py" line="93"/>
+        <source>Remind me later</source>
+        <translation>Напомнить позже</translation>
+    </message>
+    <message>
+        <location filename="../../mir_commander/ui/updates.py" line="95"/>
+        <source>Release Notes</source>
+        <translation>Список изменений</translation>
+    </message>
+    <message>
+        <location filename="../../mir_commander/ui/updates.py" line="96"/>
+        <source>Download</source>
+        <translation>Загрузить</translation>
+    </message>
+    <message>
+        <location filename="../../mir_commander/ui/updates.py" line="119"/>
+        <source>Mir Commander Update</source>
+        <translation>Обновления Mir Commander</translation>
+    </message>
+    <message>
+        <location filename="../../mir_commander/ui/updates.py" line="140"/>
+        <source>A new version is available!</source>
+        <translation>Доступна новая версия!</translation>
+    </message>
+    <message>
+        <location filename="../../mir_commander/ui/updates.py" line="141"/>
+        <source>Version {} is now available (You: v{})</source>
+        <translation>Версия {} уже доступна (текущая: v{})</translation>
+    </message>
+    <message>
+        <location filename="../../mir_commander/ui/updates.py" line="148"/>
+        <source>You are using the latest version.</source>
+        <translation>Вы используете последнюю версию.</translation>
+    </message>
+    <message>
+        <location filename="../../mir_commander/ui/updates.py" line="149"/>
+        <source>Current version: {}</source>
+        <translation>Текущая версия: {}</translation>
+    </message>
+    <message>
+        <location filename="../../mir_commander/ui/updates.py" line="153"/>
+        <source>Update check failed.</source>
+        <translation>Ошибка проверки обновления.</translation>
+    </message>
+    <message>
+        <location filename="../../mir_commander/ui/updates.py" line="154"/>
+        <source>Failed to check for updates. Check your logs for more information.</source>
+        <translation>Проверить обновления не удалось. Подробности в логах.</translation>
+    </message>
+    <message>
+        <location filename="../../mir_commander/ui/updates.py" line="160"/>
+        <source>Checking for updates...</source>
+        <translation>Проверка обновлений...</translation>
     </message>
 </context>
 <context>
@@ -65,37 +133,37 @@
 <context>
     <name>ExportFileDialog</name>
     <message>
-        <location filename="../../mir_commander/ui/export_item_dialog.py" line="47"/>
+        <location filename="../../mir_commander/ui/export_item_dialog.py" line="48"/>
         <source>Export: {}</source>
         <translation>Экспорт: {}</translation>
     </message>
     <message>
-        <location filename="../../mir_commander/ui/export_item_dialog.py" line="55"/>
+        <location filename="../../mir_commander/ui/export_item_dialog.py" line="56"/>
         <source>Format:</source>
         <translation>Формат:</translation>
     </message>
     <message>
-        <location filename="../../mir_commander/ui/export_item_dialog.py" line="64"/>
+        <location filename="../../mir_commander/ui/export_item_dialog.py" line="66"/>
         <source>Save to:</source>
         <translation>Сохранить в:</translation>
     </message>
     <message>
-        <location filename="../../mir_commander/ui/export_item_dialog.py" line="68"/>
+        <location filename="../../mir_commander/ui/export_item_dialog.py" line="70"/>
         <source>Browse...</source>
         <translation>Обзор...</translation>
     </message>
     <message>
-        <location filename="../../mir_commander/ui/export_item_dialog.py" line="74"/>
+        <location filename="../../mir_commander/ui/export_item_dialog.py" line="76"/>
         <source>Format parameters</source>
         <translation>Параметры формата</translation>
     </message>
     <message>
-        <location filename="../../mir_commander/ui/export_item_dialog.py" line="80"/>
+        <location filename="../../mir_commander/ui/export_item_dialog.py" line="82"/>
         <source>Export</source>
         <translation>Экспорт</translation>
     </message>
     <message>
-        <location filename="../../mir_commander/ui/export_item_dialog.py" line="81"/>
+        <location filename="../../mir_commander/ui/export_item_dialog.py" line="83"/>
         <source>Cancel</source>
         <translation>Отмена</translation>
     </message>
@@ -103,40 +171,99 @@
 <context>
     <name>General</name>
     <message>
-        <location filename="../../mir_commander/ui/settings/general_page.py" line="67"/>
-        <location filename="../../mir_commander/ui/settings/general_page.py" line="103"/>
+        <location filename="../../mir_commander/ui/settings/general_page.py" line="19"/>
+        <source>Interface</source>
+        <translation>Интерфейс</translation>
+    </message>
+    <message>
+        <location filename="../../mir_commander/ui/settings/general_page.py" line="26"/>
+        <source>Updates</source>
+        <translation>Обновления</translation>
+    </message>
+    <message>
+        <location filename="../../mir_commander/ui/settings/general_page.py" line="89"/>
+        <location filename="../../mir_commander/ui/settings/general_page.py" line="125"/>
         <source>System</source>
         <translation>Системный</translation>
     </message>
     <message>
-        <location filename="../../mir_commander/ui/settings/general_page.py" line="72"/>
+        <location filename="../../mir_commander/ui/settings/general_page.py" line="95"/>
         <source>Language:</source>
         <translation>Язык:</translation>
     </message>
     <message>
-        <location filename="../../mir_commander/ui/settings/general_page.py" line="74"/>
+        <location filename="../../mir_commander/ui/settings/general_page.py" line="97"/>
         <source>Language will be changed after restarting the program</source>
         <translation>Язык будет изменен после перезапуска программы</translation>
     </message>
     <message>
-        <location filename="../../mir_commander/ui/settings/general_page.py" line="98"/>
+        <location filename="../../mir_commander/ui/settings/general_page.py" line="120"/>
         <source>Font</source>
         <translation>Шрифт</translation>
     </message>
     <message>
-        <location filename="../../mir_commander/ui/settings/general_page.py" line="110"/>
+        <location filename="../../mir_commander/ui/settings/general_page.py" line="133"/>
         <source>Family:</source>
         <translation>Гарнитура:</translation>
     </message>
     <message>
-        <location filename="../../mir_commander/ui/settings/general_page.py" line="118"/>
+        <location filename="../../mir_commander/ui/settings/general_page.py" line="141"/>
         <source>Size:</source>
         <translation>Размер:</translation>
     </message>
     <message>
-        <location filename="../../mir_commander/ui/settings/general_page.py" line="121"/>
+        <location filename="../../mir_commander/ui/settings/general_page.py" line="144"/>
         <source>Font will be changed after restarting the program</source>
         <translation>Шрифт будет изменен после перезапуска программы</translation>
+    </message>
+    <message>
+        <location filename="../../mir_commander/ui/settings/general_page.py" line="198"/>
+        <source>Check for updates</source>
+        <translation>Проверять обновления</translation>
+    </message>
+    <message>
+        <location filename="../../mir_commander/ui/settings/general_page.py" line="203"/>
+        <source>Check interval:</source>
+        <translation>Интервал между проверками:</translation>
+    </message>
+    <message numerus="yes">
+        <location filename="../../mir_commander/ui/settings/general_page.py" line="207"/>
+        <location filename="../../mir_commander/ui/settings/general_page.py" line="223"/>
+        <location filename="../../mir_commander/ui/settings/general_page.py" line="242"/>
+        <source> %n hour(s)</source>
+        <translation>
+            <numerusform> час</numerusform>
+            <numerusform> часа</numerusform>
+            <numerusform> часов</numerusform>
+        </translation>
+    </message>
+</context>
+<context>
+    <name>NewVersionNotification</name>
+    <message>
+        <location filename="../../mir_commander/ui/updates.py" line="24"/>
+        <source>Download</source>
+        <translation>Загрузить</translation>
+    </message>
+    <message>
+        <location filename="../../mir_commander/ui/updates.py" line="25"/>
+        <source>Release Notes</source>
+        <translation>Список изменений</translation>
+    </message>
+    <message>
+        <location filename="../../mir_commander/ui/updates.py" line="27"/>
+        <source>Skip this version</source>
+        <translation>Пропустить эту версию</translation>
+    </message>
+    <message>
+        <location filename="../../mir_commander/ui/updates.py" line="28"/>
+        <source>Remind me later</source>
+        <translation>Напомнить позже</translation>
+    </message>
+    <message>
+        <location filename="../../mir_commander/ui/updates.py" line="53"/>
+        <source>A new version is available – v{}</source>
+        <translation>Доступна новая версия – v{}</translation>
     </message>
 </context>
 <context>
@@ -229,157 +356,157 @@
 <context>
     <name>ProjectWindow</name>
     <message>
-        <location filename="../../mir_commander/ui/project_window.py" line="86"/>
-        <source>Started {name} {version}</source>
-        <translation>Запущен {name} {version}</translation>
-    </message>
-    <message>
-        <location filename="../../mir_commander/ui/project_window.py" line="91"/>
+        <location filename="../../mir_commander/ui/project_window.py" line="100"/>
         <source>Ready</source>
         <translation>Готово</translation>
     </message>
     <message>
-        <location filename="../../mir_commander/ui/project_window.py" line="169"/>
+        <location filename="../../mir_commander/ui/project_window.py" line="184"/>
         <source>File</source>
         <translation>Файл</translation>
     </message>
     <message>
-        <location filename="../../mir_commander/ui/project_window.py" line="178"/>
+        <location filename="../../mir_commander/ui/project_window.py" line="193"/>
         <source>View</source>
         <translation>Вид</translation>
     </message>
     <message>
-        <location filename="../../mir_commander/ui/project_window.py" line="191"/>
+        <location filename="../../mir_commander/ui/project_window.py" line="206"/>
         <source>Help</source>
         <translation>Помощь</translation>
     </message>
     <message>
-        <location filename="../../mir_commander/ui/project_window.py" line="196"/>
+        <location filename="../../mir_commander/ui/project_window.py" line="212"/>
         <source>Close Project</source>
         <translation>Закрыть проект</translation>
     </message>
     <message>
-        <location filename="../../mir_commander/ui/project_window.py" line="215"/>
+        <location filename="../../mir_commander/ui/project_window.py" line="231"/>
         <source>Quit</source>
         <translation>Выход</translation>
     </message>
     <message>
-        <location filename="../../mir_commander/ui/project_window.py" line="228"/>
+        <location filename="../../mir_commander/ui/project_window.py" line="244"/>
+        <source>Check for Updates</source>
+        <translation>Проверить обновления</translation>
+    </message>
+    <message>
+        <location filename="../../mir_commander/ui/project_window.py" line="250"/>
         <source>Import File...</source>
         <translation>Импорт файла...</translation>
     </message>
     <message>
-        <location filename="../../mir_commander/ui/project_window.py" line="234"/>
+        <location filename="../../mir_commander/ui/project_window.py" line="256"/>
         <source>Close</source>
         <translation>Закрыть</translation>
     </message>
     <message>
-        <location filename="../../mir_commander/ui/project_window.py" line="238"/>
+        <location filename="../../mir_commander/ui/project_window.py" line="260"/>
         <source>Close All</source>
         <translation>Закрыть все</translation>
     </message>
     <message>
-        <location filename="../../mir_commander/ui/project_window.py" line="242"/>
+        <location filename="../../mir_commander/ui/project_window.py" line="264"/>
         <source>Tile</source>
         <translation>Замостить</translation>
     </message>
     <message>
-        <location filename="../../mir_commander/ui/project_window.py" line="245"/>
+        <location filename="../../mir_commander/ui/project_window.py" line="267"/>
         <source>Cascade</source>
         <translation>Каскадировать</translation>
     </message>
     <message>
-        <location filename="../../mir_commander/ui/project_window.py" line="249"/>
+        <location filename="../../mir_commander/ui/project_window.py" line="271"/>
         <source>Next</source>
         <translation>Следующее</translation>
     </message>
     <message>
-        <location filename="../../mir_commander/ui/project_window.py" line="257"/>
+        <location filename="../../mir_commander/ui/project_window.py" line="279"/>
         <source>Previous</source>
         <translation>Предыдущее</translation>
     </message>
     <message>
-        <location filename="../../mir_commander/ui/project_window.py" line="234"/>
+        <location filename="../../mir_commander/ui/project_window.py" line="256"/>
         <source>Close the active window</source>
         <translation>Закрыть активное окно</translation>
     </message>
     <message>
-        <location filename="../../mir_commander/ui/project_window.py" line="185"/>
+        <location filename="../../mir_commander/ui/project_window.py" line="200"/>
         <source>Window</source>
         <translation>Окно</translation>
     </message>
     <message>
-        <location filename="../../mir_commander/ui/project_window.py" line="201"/>
+        <location filename="../../mir_commander/ui/project_window.py" line="217"/>
         <source>Preferences...</source>
         <translation>Настройки...</translation>
     </message>
     <message>
-        <location filename="../../mir_commander/ui/project_window.py" line="222"/>
+        <location filename="../../mir_commander/ui/project_window.py" line="238"/>
         <source>About Mir Commander</source>
         <translation>О Mir Commander</translation>
     </message>
     <message>
-        <location filename="../../mir_commander/ui/project_window.py" line="239"/>
+        <location filename="../../mir_commander/ui/project_window.py" line="261"/>
         <source>Close all the windows</source>
         <translation>Закрыть все окна</translation>
     </message>
     <message>
-        <location filename="../../mir_commander/ui/project_window.py" line="242"/>
+        <location filename="../../mir_commander/ui/project_window.py" line="264"/>
         <source>Tile the windows</source>
         <translation>Расположить все окна замощением</translation>
     </message>
     <message>
-        <location filename="../../mir_commander/ui/project_window.py" line="245"/>
+        <location filename="../../mir_commander/ui/project_window.py" line="267"/>
         <source>Cascade the windows</source>
         <translation>Расположить все окна каскадно</translation>
     </message>
     <message>
-        <location filename="../../mir_commander/ui/project_window.py" line="252"/>
+        <location filename="../../mir_commander/ui/project_window.py" line="274"/>
         <source>Move the focus to the next window</source>
         <translation>Переместить фокус на следующее окно</translation>
     </message>
     <message>
-        <location filename="../../mir_commander/ui/project_window.py" line="260"/>
+        <location filename="../../mir_commander/ui/project_window.py" line="282"/>
         <source>Move the focus to the previous window</source>
         <translation>Переместить фокус на предыдущее окно</translation>
     </message>
     <message>
-        <location filename="../../mir_commander/ui/project_window.py" line="339"/>
+        <location filename="../../mir_commander/ui/project_window.py" line="366"/>
         <source>Import File</source>
         <translation>Импорт файла</translation>
     </message>
     <message>
-        <location filename="../../mir_commander/ui/project_window.py" line="341"/>
+        <location filename="../../mir_commander/ui/project_window.py" line="368"/>
         <source>All files (*)</source>
         <translation>Все файлы (*)</translation>
     </message>
     <message>
-        <location filename="../../mir_commander/ui/project_window.py" line="110"/>
+        <location filename="../../mir_commander/ui/project_window.py" line="125"/>
         <source>File imported successfully</source>
         <translation>Файл успешно импортирован</translation>
     </message>
     <message>
-        <location filename="../../mir_commander/ui/project_window.py" line="114"/>
+        <location filename="../../mir_commander/ui/project_window.py" line="129"/>
         <source>Error importing file {file_path}: {e}</source>
         <translation>Ошибка импорта файла {file_path}: {e}</translation>
     </message>
     <message>
-        <location filename="../../mir_commander/ui/project_window.py" line="115"/>
+        <location filename="../../mir_commander/ui/project_window.py" line="130"/>
         <source>Failed to import file</source>
         <translation>Не удалось импортировать файл</translation>
     </message>
     <message>
-        <location filename="../../mir_commander/ui/project_window.py" line="356"/>
+        <location filename="../../mir_commander/ui/project_window.py" line="383"/>
         <source>File exported successfully</source>
         <translation>Файл успешно экспортирован</translation>
     </message>
     <message>
-        <location filename="../../mir_commander/ui/project_window.py" line="359"/>
+        <location filename="../../mir_commander/ui/project_window.py" line="386"/>
         <source>Failed to export file: {}</source>
         <translation>Не удалось экспортировать файл: {}</translation>
     </message>
     <message>
-        <location filename="../../mir_commander/ui/project_window.py" line="360"/>
+        <location filename="../../mir_commander/ui/project_window.py" line="387"/>
         <source>Failed to export file</source>
         <translation>Не удалось экспортировать файл</translation>
     </message>
@@ -435,12 +562,12 @@
         <translation>Ок</translation>
     </message>
     <message>
-        <location filename="../../mir_commander/ui/settings/settings_dialog.py" line="109"/>
+        <location filename="../../mir_commander/ui/settings/settings_dialog.py" line="110"/>
         <source>Project</source>
         <translation>Проект</translation>
     </message>
     <message>
-        <location filename="../../mir_commander/ui/settings/settings_dialog.py" line="110"/>
+        <location filename="../../mir_commander/ui/settings/settings_dialog.py" line="109"/>
         <source>General</source>
         <translation>Основные</translation>
     </message>